requirements.txt
fastapi==0.115.6
fitz==0.0.1.dev2
img2table==1.4.0
langchain==0.3.13
langchain_chroma==0.1.4
langchain_community==0.3.13
langchain_core==0.3.28
langchain_experimental==0.3.4
langchain_google_genai==2.0.7
llama_parse==0.5.19
modules==1.0.0
nest_asyncio==1.5.6
nest_asyncio==1.6.0
opencv_contrib_python==4.10.0.84
opencv_python_headless==4.10.0.84
pydantic==1.10.19
pydantic==1.10.8
pymupdf4llm==0.0.17
python-dotenv==1.0.1
uuid6==2024.7.10
weaviate_client==4.10.2
